README.md
The program analyzes the machine code and restores all the functions of the program, combining them into logical modules and visualizing them in 2D space. User-defined functions are allocated into clusters according to the k-means algorithm. The program is used to depict how the system functions as a whole and at the individual level, simplifying the development of complex applications or software systems by stimulating the reuse of software, its development, analysis and management. The program is designed to describe approaches to defining and compiling security requirements for software modules that have different quality of service requirements and differ in incompatible security restrictions.